part6 27 imlement some updates and fix some bugs

+++ src/com/bookstore/service/CategoryServices.java
@@ -7,6 +7,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.bookstore.dao.BookDAO;
 import com.bookstore.dao.CategoryDAO;
 import com.bookstore.entity.Category;
 
@@ -94,9 +95,18 @@ public class CategoryServices {
 
 	public void deleteCategory() throws ServletException, IOException {
 		int categoryId = Integer.parseInt(request.getParameter("id"));
-		categoryDAO.delete(categoryId);
+		BookDAO bookDAO = new BookDAO();
+		long numberOfBooks = bookDAO.countByCategory(categoryId);
+		String message;
 		
-		String message = "The category with ID " + categoryId + " has been removed successfully";
+		if (numberOfBooks > 0) {
+			message = "Could not delete the category (ID: %d) because it currently contains some books";
+			message = String.format(message, numberOfBooks);
+		} else {
+			categoryDAO.delete(categoryId);
+			message = "The category with ID " + categoryId + " has been removed successfully";
+		}
+		
 		listCategory(message);	
 	}
